refactor: split large components
