+++ src/components/ChantCard.jsx
@@ -1,0 +1,91 @@
+import React from 'react';
+import { Play, Share2, Plus, Heart } from 'lucide-react';
+import { getTeamInfo, getPositionKorean } from '../utils/team';
+
+const ChantCard = ({
+  chant,
+  playerSongs,
+  setCurrentPlayer,
+  setPlaySource,
+  setShowPlayer,
+  toggleLike,
+  likedSongs
+}) => (
+  <div className="bg-gradient-to-br from-white to-gray-50 rounded-2xl border border-gray-200 p-4 hover:shadow-lg hover:border-gray-300 transition-all duration-200">
+    <div className="flex items-start justify-between mb-4">
+      <div className="flex-1">
+        <h3 className="font-bold text-lg text-gray-900">{chant.playerName}</h3>
+        <div className="flex flex-wrap items-center gap-2 text-sm text-gray-600 mt-1">
+          <div className="flex items-center gap-1">
+            {getTeamInfo(chant.team).logo && (
+              <img
+                src={getTeamInfo(chant.team).logo}
+                alt={chant.team}
+                className="w-4 h-4 object-contain"
+              />
+            )}
+            <span className="font-medium px-2 py-1 rounded-full text-xs" style={{
+              backgroundColor: `${getTeamInfo(chant.team).color}15`,
+              color: getTeamInfo(chant.team).color
+            }}>
+              {chant.team}
+            </span>
+          </div>
+          {chant.position && (
+            <>
+              <span className="text-gray-300">•</span>
+              <span className="bg-gray-100 px-2 py-1 rounded-full text-xs">
+                {getPositionKorean(chant.position)}
+              </span>
+            </>
+          )}
+          <span className="text-gray-300">•</span>
+          <div className="flex items-center gap-1">
+            <Heart className="w-3 h-3 text-red-400" />
+            <span className="text-xs text-gray-500">{chant.likes.toLocaleString()}</span>
+          </div>
+        </div>
+      </div>
+      <button
+        onClick={e => {
+          e.stopPropagation();
+          toggleLike(chant.id);
+        }}
+        className="p-2 rounded-full hover:bg-gray-100 transition-colors"
+      >
+        <Heart
+          className={`w-5 h-5 transition-all ${
+            likedSongs.has(chant.id)
+              ? 'text-red-500 fill-red-500 scale-110'
+              : 'text-gray-300 hover:text-red-400'
+          }`}
+        />
+      </button>
+    </div>
+
+    <div className="flex items-center gap-2">
+      <button
+        className="flex-1 bg-gradient-to-r from-blue-500 to-indigo-600 text-white py-3 px-4 rounded-xl hover:from-blue-600 hover:to-indigo-700 transition-all flex items-center justify-center gap-2 shadow-md hover:shadow-lg"
+        onClick={() => {
+          const playerIndex = playerSongs.findIndex(c => c.id === chant.id);
+          if (playerIndex !== -1) {
+            setCurrentPlayer(playerIndex);
+            setPlaySource('explore');
+            setShowPlayer(true);
+          }
+        }}
+      >
+        <Play className="w-4 h-4" />
+        재생
+      </button>
+      <button className="p-3 border border-gray-200 rounded-xl hover:bg-gray-50 hover:border-gray-300 transition-all">
+        <Share2 className="w-4 h-4 text-gray-600" />
+      </button>
+      <button className="p-3 border border-gray-200 rounded-xl hover:bg-gray-50 hover:border-gray-300 transition-all">
+        <Plus className="w-4 h-4 text-gray-600" />
+      </button>
+    </div>
+  </div>
+);
+
+export default ChantCard;
